Add IO::AwsS3 to upload traces to AWS S3
We have a plan to save runners' traces to AWS S3.
In order to do that, I added `IO::AwsS3` class.
It behaves like `IO` objects, but it has `#finalize!` method,
which uploads all traces to S3.
Before `#finalize!` is called, an `IO::AwsS3` instance writes
traces to a temporary file as well as the passed STDOUT.
I chose a temporary file as the temporary store instead of a memory
store like `StringIO` because I can't estimate the usage of memory.

The feature of uploading to S3 is currently optional.
If you want to use this feature, `--output` option is required.
Of course, you have to prepare an AWS S3 bucket and
allow `Runners` to upload files.