Show deprecation warning for too old ESLint
This aims to promote upgrading for users which are using too old ESLint versions (The latest version is `6.4.0`).
When I examine our statistics, the oldest ESLint used on Sider is `3.19.1` (the latest of 3.x series).
The next is `4.19.1` (the latest of 4.x series).

I think that such work is common for other tools, so I add `Processor#add_warning_if_deprecated_version` method.