Flush content whenever its size reaches the buffer max limit
Currently, the trace of the analysis is flushed at the end of the
analysis. However, I want to see the on-going trace log even though the
analysis has not yet finished. Also, I want to know whether the analysis
has started or not, so at the beginning of the analysis, I inserted the
`io.flush!` procedure.