Add a feature to mask secure strings on TraceWriter
We are planning to fetch source code via Git. To get source code
from a private repository, we must set a token and perform command like
this:

```shell-session
$ git config http.extraheader "Authorization: Basic token"
```

I'm going to run the command with `Shell#capture3!`. However, this
command writes its passed command line via `TraceWriter`. So, we have to
mask the secure string and prevent it from being output in traces.

Secure strings also should be masked on stdout or stderr, so I
implemented the feature on `TraceWriter` instead of `Shell`.