Add Runners::Config, which is responsible for sider.yml
This PR introduces `Runners::Config` which is responsible for our
configuration file `sider.yml`. Currently, this file is parsed in each
processor, but as it's told in #750, unknown keys or other top-level
attributes are not validated. This commit just adds `Runners::Config`,
so I will make Runners use the class in the subsequent commit.