Pass :config to Processor (#779)
This is for the To-Do "Pass config to Processor#initialize from Harness" of #750.

Note that it contains the breaking change; the field name "golangci-lint" in `sider.yml` has changed to "golangci_lint" for consistency with other runners. However, we haven't yet released the new version, which includes "runner_golangci_lint," so there are few impacts.