add type signature of time class from rbs v1.0.0 (#1962)
This is required to use class Time in a runner processer with type checking by RBS.

NOTE: It is better to upgrade rbs gem ideally. But we cannot do that now because steep has not supported them yet (#1905). We should remove this code when we are ready to update rbs gem.